+++ docs/changelog.rst
@@ -14,6 +14,7 @@ Changelog
 * Fixed an issue that prevented the Editor View to be restarted in some circumnstances.
 * Fixed the translation of multiple unaligned objects when snapping was enabled.
 * Added new Crown logo.
+* Added icons to the Level Tree View.
 
 **Runtime**